Blog Post 3
So, I added all my saved things from my github account in my last
commit, and two blog posts were in it. The second one had a playful
simplicity about its surface that made it entertaining. The third one,
posted with this comment, was intriguing bcause of how the shapes on the
home page bounced freely off the sides of the window, just like sound
waves.